fix: API consistency, move tests, fix test names, total extension overflow check
1. set_extension() now rejects 0-byte payloads for one-byte extensions,
   matching marshal_to() validation (RFC 8285 §4.2).
2. Moved header tests to header/header_test.rs (header.rs -> header/mod.rs).
3. Renamed misleading test_h265_flush_single_nalu_passthrough to
   test_h265_flush_single_oversized_nalu_fu_fragmentation.
4. Added defense-in-depth comment on unreachable u16::MAX branch in
   flush_aggregation_buffer.
5. Added ExtensionPayloadTotalOverflow error and overflow check before
   u16 cast in marshal_to().
